requirements.txt
$requirements = @"
Flask==2.3.0
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
"@

$requirements | Out-File -Encoding UTF8 requirements.txt -Force
